+++ lib/rubocop.rb
@@ -75,6 +75,7 @@ require 'rubocop/cop/leading_comment_space'
 require 'rubocop/cop/colon_method_call'
 require 'rubocop/cop/not'
 require 'rubocop/cop/eval'
+require 'rubocop/cop/access_control'
 
 require 'rubocop/report/report'
 require 'rubocop/report/plain_text'