+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ### New features
 
 * New cop `ConstantName` checks for constant which are not using `SCREAMING_SNAKE_CASE`.
+* New cop `AccessControl` check private/protected indentation and surrounding blank lines.
 
 ## 0.7.2 (05/13/2013)